refactor(tui): extract view shell rendering component
